+++ app/src/main/java/elektrogo/front/MainActivity.kt
@@ -9,8 +9,8 @@ import com.google.android.material.bottomnavigation.BottomNavigationView
 import androidx.fragment.app.Fragment
 import com.google.android.material.navigation.NavigationBarView
 import elektrogo.front.databinding.ActivityMainBinding
-import elektrogo.front.ui.MapsFragment
-import elektrogo.front.ui.ProfileFragment
+import elektrogo.front.ui.map.MapsFragment
+import elektrogo.front.ui.profile.ProfileFragment
 
 class MainActivity : AppCompatActivity() {
