+++ app/src/main/java/elektrogo/front/ui/profile/ProfileFragment.kt
@@ -1,4 +1,4 @@
-package elektrogo.front.ui
+package elektrogo.front.ui.profile
 
 import androidx.lifecycle.ViewModelProvider
 import android.content.Intent
@@ -6,11 +6,9 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
-import elektrogo.front.AddVehicle
+import elektrogo.front.ui.vehicleAdd.AddVehicle
 import elektrogo.front.databinding.ProfileFragmentBinding
-import elektrogo.front.ui.ProfileViewModel
 
 class ProfileFragment : Fragment() {
