+++ app/src/main/java/elektrogo/front/ui/vehicleList/VehicleListItemViewModel.kt
@@ -1,8 +1,6 @@
 package elektrogo.front.ui.vehicleList
 
 import androidx.lifecycle.ViewModel
-import elektrogo.front.FrontendController
-import kotlinx.coroutines.runBlocking
 
 class VehicleListItemViewModel : ViewModel() {
