+++ app/src/main/res/layout/activity_add_vehicle.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddVehicle">
+    tools:context=".ui.vehicleAdd.AddVehicle">
 
     <ScrollView
         android:layout_width="match_parent"
